'utils' > 'util', check for entry length, doc spelling
