+++ genestack_client/__init__.py
@@ -21,4 +21,4 @@ from files_util import FilesUtil, SpecialFolders
 from task_log_viewer import TaskLogViewer
 from utils import get_connection, make_connection_parser, get_user
 from cla import *
-from dictionary_utils import DictionaryUtils
+from dictionary_util import DictionaryUtil
